+++ routes/campuses.mjs
@@ -15,6 +15,7 @@ import {
 } from '../models/rights';
 import contentNegociation from '../middlewares/content-negociation';
 import maskOutput from '../middlewares/mask-output';
+import { emitDriverConnection } from '../middlewares/drivers-socket-status';
 
 const BASIC_OUTPUT_MASK = '_id,id,name,location(coordinates),phone(everybody),defaultReservationScope';
 const router = generateCRUD(Campus, {
@@ -46,6 +47,7 @@ const router = generateCRUD(Campus, {
     lean: false,
     right: [CAN_GET_CAMPUS_BASIC, CAN_GET_CAMPUS],
     middlewares: [
+      emitDriverConnection,
       async (ctx, next) => {
         await next();
         if (!ctx.may(CAN_GET_CAMPUS)) {